+++ spec/lib/aixm/document_spec.rb
@@ -245,6 +245,7 @@ describe AIXM::Document do
             <valAuwWeight>30</valAuwWeight>
             <uomAuwWeight>T</uomAuwWeight>
             <codeSts>CLSD</codeSts>
+            <txtMarking>Standard marking</txtMarking>
             <txtRmk>Markings eroded</txtRmk>
           </Rwy>
           <Rdn>
@@ -1063,6 +1064,7 @@ describe AIXM::Document do
             <valAuwWeight>30</valAuwWeight>
             <uomAuwWeight>T</uomAuwWeight>
             <codeSts>CLSD</codeSts>
+            <txtMarking>Standard marking</txtMarking>
             <txtRmk>Markings eroded</txtRmk>
           </Rwy>
           <Rdn>
@@ -1959,6 +1961,7 @@ describe AIXM::Document do
             <valAuwWeight>30</valAuwWeight>
             <uomAuwWeight>T</uomAuwWeight>
             <codeSts>CLSD</codeSts>
+            <txtMarking>Standard marking</txtMarking>
             <txtRmk>Markings eroded</txtRmk>
           </Rwy>
           <Rdn>